Docs, linter, formatting
