+++ app/gui/view/component-browser/searcher-list-panel/src/lib.rs
@@ -89,6 +89,9 @@ use searcher_theme::list_panel as list_panel_theme;
 pub mod column_grid;
 pub mod layout;
 pub mod layouting;
+
+
+
 mod navigator;
 
 pub use column_grid::LabeledAnyModelProvider;
