+++ lib/rust/ensogl/component/grid-view/src/header.rs
@@ -18,6 +18,7 @@ use ensogl_core::display::scene::Layer;
 use ensogl_scroll_area::Viewport;
 
 
+
 // ===========
 // === FRP ===
 // ===========
